+++ app/helpers/ndr_ui/bootstrap_helper.rb
@@ -243,6 +243,34 @@ module NdrUi
       end
     end
 
+    # Identical signature to form_with, but uses NdrUi::BootstrapBuilder.
+    # See ActionView::Helpers::FormHelper for details
+    def bootstrap_form_with(**options, &block)
+      options[:html] ||= {}
+      options[:builder] = NdrUi::BootstrapBuilder
+      horizontal = options.delete(:horizontal)
+
+      # :horizontal
+      if horizontal
+        # set the form html class for horizontal bootstrap forms
+        options[:html][:class] ||= ''
+        classes = (options[:html][:class].split(' ') << 'form-horizontal').uniq.join(' ')
+        options[:html][:class] = classes
+      end
+
+      # We switch autocomplete off by default
+      raise 'autocomplete should be defined an html option' if options[:autocomplete]
+      options[:html][:autocomplete] ||= 'off'
+
+      form_with(**options) do |form|
+        # Put the form builder into horizontal mode (if necessary)
+        form.horizontal_mode = horizontal if horizontal
+
+        # yield to the provided form block
+        block.call(form)
+      end
+    end
+
     # TODO: bootstrap_pagination_tag(*args, &block)
 
     # Creates a Boostrap control group for button.